Changes - About us new page and images - Location map responsiveness
